+++ src/azure-cli-core/azure/cli/core/extensions/transform.py
@@ -16,7 +16,7 @@ def register(application):
 def _parse_id(strid):
     parsed = {}
     parts = re.split('/', strid)
-    if parts[3] != 'resourceGroups':
+    if parts[3].lower() != 'resourcegroups':
         raise KeyError()
 
     parsed['resource-group'] = parts[4]