+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/tests/test_custom.py
@@ -13,7 +13,9 @@ import yaml
 
 from msrestazure.azure_exceptions import CloudError
 
-from azure.cli.command_modules.acs.custom import (merge_kubernetes_configurations,
+from azure.cli.command_modules.acs._params import (regionsInPreview,
+                                                   regionsInProd)
+from azure.cli.command_modules.acs.custom import (merge_kubernetes_configurations, list_acs_locations,
                                                   _acs_browse_internal, _add_role_assignment, _get_default_dns_prefix)
 from azure.mgmt.containerservice.models import (ContainerServiceOrchestratorTypes,
                                                 ContainerService,
@@ -22,6 +24,13 @@ from azure.cli.core.util import CLIError
 
 
 class AcsCustomCommandTest(unittest.TestCase):
+    def test_list_acs_locations(self):
+        regions = list_acs_locations()
+        prodregions = regions["productionRegions"]
+        previewregions = regions["previewRegions"]
+        self.assertListEqual(prodregions, regionsInProd, "Production regions doesn't match")
+        self.assertListEqual(previewregions, regionsInPreview, "Preview regions doesn't match")
+
     def test_get_default_dns_prefix(self):
         name = 'test5678910'
         resource_group_name = 'resource_group_with_underscore'
@@ -40,7 +49,7 @@ class AcsCustomCommandTest(unittest.TestCase):
 
         with mock.patch(
                 'azure.cli.command_modules.acs.custom.create_role_assignment') as create_role_assignment:
-            ok = _add_role_assignment(role, sp, delay=0, output=False)
+            ok = _add_role_assignment(role, sp, delay=0)
             create_role_assignment.assert_called_with(role, sp)
             self.assertTrue(ok, 'Expected _add_role_assignment to succeed')
 
@@ -50,13 +59,13 @@ class AcsCustomCommandTest(unittest.TestCase):
 
         with mock.patch(
                 'azure.cli.command_modules.acs.custom.create_role_assignment') as create_role_assignment:
-            resp = mock.Mock()
+            resp = mock.create_autospec(requests.Response)
             resp.status_code = 409
-            resp.content = 'Conflict'
+            resp.text = 'Conflict'
             err = CloudError(resp)
             err.message = 'The role assignment already exists.'
             create_role_assignment.side_effect = err
-            ok = _add_role_assignment(role, sp, delay=0, output=False)
+            ok = _add_role_assignment(role, sp, delay=0)
 
             create_role_assignment.assert_called_with(role, sp)
             self.assertTrue(ok, 'Expected _add_role_assignment to succeed')
@@ -67,13 +76,13 @@ class AcsCustomCommandTest(unittest.TestCase):
 
         with mock.patch(
                 'azure.cli.command_modules.acs.custom.create_role_assignment') as create_role_assignment:
-            resp = mock.Mock()
+            resp = mock.create_autospec(requests.Response)
             resp.status_code = 500
-            resp.content = 'Internal Error'
+            resp.text = 'Internal Error'
             err = CloudError(resp)
             err.message = 'Internal Error'
             create_role_assignment.side_effect = err
-            ok = _add_role_assignment(role, sp, delay=0, output=False)
+            ok = _add_role_assignment(role, sp, delay=0)
 
             create_role_assignment.assert_called_with(role, sp)
             self.assertFalse(ok, 'Expected _add_role_assignment to fail')