+++ src/command_modules/azure-cli-redis/azure/cli/command_modules/redis/custom.py
@@ -23,31 +23,29 @@ def cli_redis_import_method(client, resource_group_name, name, file_format, file
 
 
 def cli_redis_update_settings(client, resource_group_name, name, redis_configuration):
-    from azure.mgmt.redis.models import RedisCreateOrUpdateParameters
+    from azure.mgmt.redis.models import RedisUpdateParameters
     logger.warning('This command is getting deprecated. Please use "redis update" command')
 
     existing = client.get(resource_group_name, name)
     existing.redis_configuration.update(redis_configuration)
 
     # Due to swagger/mgmt SDK quirkiness, we have to manually copy over
-    # the resource retrieved to a create_or_update_parameters object
-    update_params = RedisCreateOrUpdateParameters(
-        existing.location,
-        existing.sku,
-        existing.tags,
-        existing.redis_version,
+    # the resource retrieved to a create_parameters object
+    update_params = RedisUpdateParameters(
         existing.redis_configuration,
         existing.enable_non_ssl_port,
         existing.tenant_settings,
         existing.shard_count,
         existing.subnet_id,
         existing.static_ip,
+        existing.sku,
+        existing.tags
     )
-    return client.create_or_update(resource_group_name, name, parameters=update_params)
+    return client.update(resource_group_name, name, parameters=update_params)
 
 
 def cli_redis_update(instance, sku=None, vm_size=None):
-    from azure.mgmt.redis.models import RedisCreateOrUpdateParameters
+    from azure.mgmt.redis.models import RedisUpdateParameters
     if sku is not None:
         instance.sku.name = sku
 
@@ -55,17 +53,15 @@ def cli_redis_update(instance, sku=None, vm_size=None):
         instance.sku.family = vm_size[0]
         instance.sku.capacity = vm_size[1:]
 
-    update_params = RedisCreateOrUpdateParameters(
-        instance.location,
-        instance.sku,
-        instance.tags,
-        instance.redis_version,
+    update_params = RedisUpdateParameters(
         instance.redis_configuration,
         instance.enable_non_ssl_port,
         instance.tenant_settings,
         instance.shard_count,
         instance.subnet_id,
         instance.static_ip,
+        instance.sku,
+        instance.tags
     )
 
     return update_params
@@ -103,12 +99,11 @@ def cli_redis_create(client,
     :param subnet_id: The full resource ID of a subnet in a virtual network to deploy the redis cache in. Example format /subscriptions/{subid}/resourceGroups/{resourceGroupName}/Microsoft.{Network|ClassicNetwork}/VirtualNetworks/vnet1/subnets/subnet1
     :param static_ip: Required when deploying a redis cache inside an existing Azure Virtual Network.
     """
-    from azure.mgmt.redis.models import RedisCreateOrUpdateParameters, Sku
-    params = RedisCreateOrUpdateParameters(
+    from azure.mgmt.redis.models import RedisCreateParameters, Sku
+    params = RedisCreateParameters(
         location,
         Sku(sku, vm_size[0], vm_size[1:]),
         tags,
-        None,  # Version is deprecated and ignored
         redis_configuration,
         enable_non_ssl_port,
         tenant_settings,
@@ -116,4 +111,4 @@ def cli_redis_create(client,
         subnet_id,
         static_ip)
 
-    return client.create_or_update(resource_group_name, name, params)
+    return client.create(resource_group_name, name, params)